styles advSec  / Container fixed / Container moved to SharedLayout
