+++ src/components/AdvantagesSection/AdvantagesSection.module.css
@@ -1,5 +1,5 @@
 .advantagesSection {
-  width: 343px;
+  min-width: 288px;
   height: 363px;
   border-radius: 30px;
   background-color: var(--light);
