+++ src/components/Container/Container.module.css
@@ -1,8 +1,9 @@
 .container {
+  min-width: 320px;
   max-width: 375px;
   margin: 0 auto;
   padding: 0 16px;
-  
+
 }
 
 @media screen and (min-width: 768px) {
