--- src/components/SharedLayout/SharedLayout.module.css
@@ -1,11 +1,0 @@
-.SharedLayout {
-    margin: 0 auto;
-    padding: 16px;
-    max-width: 100%;
-}
-
-@media screen and (min-width: 768px) {
-    .SharedLayout {
-        padding: 32px;
-    }
-}
