+++ src/pages/SignInPage/SignInPage.jsx
@@ -1,12 +1,10 @@
 import css from './SignInPage.module.css';
 import { AdvantagesSection } from '../../components/AdvantagesSection/AdvantagesSection';
 import SignInForm from '../../components/SignInForm/SignInForm';
-import Container from '../../components/Container/Container';
 import Page from '../../components/Page/Page';
 
 const SignInPage = () => {
   return (
-    <Container>
       <Page>
         <title>Sign In Page</title>
         <SignInForm />
@@ -14,7 +12,6 @@ const SignInPage = () => {
           <AdvantagesSection />
         </div>
       </Page>
-    </Container>
   );
 };
