+++ src/pages/SignUpPage/SignUpPage.jsx
@@ -1,12 +1,10 @@
 import css from './SignUpPage.module.css';
 import { AdvantagesSection } from '../../components/AdvantagesSection/AdvantagesSection';
-import Container from '../../components/Container/Container';
 import Page from '../../components/Page/Page';
 import SignUpForm from '../../components/SignUpForm/SignUpForm';
 
 const SignUpPage = () => {
   return (
-    <Container>
       <Page>
         <title>Sign Up Page</title>
         <SignUpForm />
@@ -14,7 +12,6 @@ const SignUpPage = () => {
           <AdvantagesSection />
         </div>
       </Page>
-    </Container>
   );
 };
